Change textSize field type from int to float

+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/customcontrols/ControlData.java
@@ -85,7 +85,7 @@ public class ControlData {
     }
 
     public String name;
-    public int textSize;        //In Dp
+    public float textSize;        //In Dp
     private float width;         //Dp instead of Px now
     private float height;        //Dp instead of Px now
     public int[] keycodes;      //Should store up to 4 keys
@@ -138,10 +138,10 @@ public class ControlData {
     }
 
     public ControlData(String name, int[] keycodes, String dynamicX, String dynamicY, float width, float height, boolean isToggle){
-        this(name, 14, keycodes, dynamicX, dynamicY, width, height, isToggle, 1,0x4D000000, 0xFFFFFFFF,0,0);
+        this(name, 14, keycodes, dynamicX, dynamicY, width, height, isToggle, 1, 0x4D000000, 0xFFFFFFFF, 0, 0);
     }
 
-    public ControlData(String name, int textSize, int[] keycodes, String dynamicX, String dynamicY, float width, float height, boolean isToggle, float opacity, int bgColor, int strokeColor, int strokeWidth, float cornerRadius) {
+    public ControlData(String name, float textSize, int[] keycodes, String dynamicX, String dynamicY, float width, float height, boolean isToggle, float opacity, int bgColor, int strokeColor, int strokeWidth, float cornerRadius) {
         this.name = name;
         this.textSize = textSize;
         this.keycodes = inflateKeycodeArray(keycodes);
@@ -221,7 +221,7 @@ public class ControlData {
         return Tools.dpToPx(height);
     }
 
-    public int getTextSize(){
+    public float getTextSize(){
         return textSize;
     }
 
@@ -234,7 +234,7 @@ public class ControlData {
         height = Tools.pxToDp(heightInPx);
     }
 
-    public void setTextSize(int sizeInDp){
+    public void setTextSize(float sizeInDp){
         textSize = sizeInDp;
     }
